+++ gtime/forecasting/gar.py
@@ -9,7 +9,7 @@ from sklearn.multioutput import (
 import numpy as np
 from sklearn.utils.validation import check_is_fitted
 
-from .multi_output import MultiFeatureMultiOutputRegressor
+from gtime.regressors.multi_output import MultiFeatureMultiOutputRegressor
 
 
 class GAR(MultiOutputRegressor):
